+++ realm/realm-library/src/androidTest/java/io/realm/RealmObjectTests.java
@@ -26,6 +26,7 @@ import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 
 import java.io.FileNotFoundException;
+import java.lang.ref.WeakReference;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.concurrent.Callable;
@@ -37,6 +38,7 @@ import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicReference;
 
 import io.realm.entities.AllTypes;
+import io.realm.entities.AllTypesPrimaryKey;
 import io.realm.entities.ConflictingFieldName;
 import io.realm.entities.CustomMethods;
 import io.realm.entities.CyclicType;
@@ -961,7 +963,7 @@ public class RealmObjectTests {
     }
 
     @Test
-    public void IsValid_deletedObject() {
+    public void isValid_deletedObject() {
         realm.beginTransaction();
         AllTypes allTypes = realm.createObject(AllTypes.class);
         assertTrue(allTypes.isValid());
@@ -1666,6 +1668,109 @@ public class RealmObjectTests {
             fail("Failed to remove null listener.");
         } catch (IllegalArgumentException ignore) {
             looperThread.testComplete();
+        }
+    }
+
+    // Bug https://github.com/realm/realm-java/issues/2569
+    @Test
+    @RunTestInLooperThread
+    public void addChangeListener_returnedObjectOfCopyToRealmOrUpdate() {
+        looperThread.realm.beginTransaction();
+        looperThread.realm.createObject(AllTypesPrimaryKey.class, 1);
+
+        AllTypesPrimaryKey allTypesPrimaryKey = new AllTypesPrimaryKey();
+        allTypesPrimaryKey.setColumnLong(1);
+        allTypesPrimaryKey.setColumnFloat(42f);
+        allTypesPrimaryKey = looperThread.realm.copyToRealmOrUpdate(allTypesPrimaryKey);
+        looperThread.realm.commitTransaction();
+
+        allTypesPrimaryKey.addChangeListener(new RealmChangeListener<AllTypesPrimaryKey>() {
+            @Override
+            public void onChange(AllTypesPrimaryKey element) {
+                assertEquals(42.0f, element.getColumnFloat(), 0f);
+                looperThread.testComplete();
+            }
+        });
+    }
+
+    // The object should be added to HandlerController.realmObjects only when the first time addListener called.
+    @Test
+    @RunTestInLooperThread
+    public void addChangeListener_shouldAddTheObjectToHandlerRealmObjects() {
+        looperThread.realm.beginTransaction();
+        AllTypesPrimaryKey allTypesPrimaryKey = looperThread.realm.createObject(AllTypesPrimaryKey.class, 1);
+        looperThread.realm.commitTransaction();
+        assertTrue(looperThread.realm.handlerController.realmObjects.isEmpty());
+
+        allTypesPrimaryKey.addChangeListener(new RealmChangeListener<RealmModel>() {
+            @Override
+            public void onChange(RealmModel element) {
+            }
+        });
+
+        assertEquals(1, looperThread.realm.handlerController.realmObjects.size());
+        for (WeakReference<RealmObjectProxy> ref : looperThread.realm.handlerController.realmObjects.keySet()) {
+            assertTrue(ref.get() == allTypesPrimaryKey);
+        }
+
+        looperThread.testComplete();
+    }
+
+    // The object should be added to HandlerController.realmObjects only once.
+    @Test
+    @RunTestInLooperThread
+    public void addChangeListener_shouldNotAddDupEntriesToHandlerRealmObjects() {
+        looperThread.realm.beginTransaction();
+        AllTypesPrimaryKey allTypesPrimaryKey = looperThread.realm.createObject(AllTypesPrimaryKey.class, 1);
+        looperThread.realm.commitTransaction();
+        for (WeakReference<RealmObjectProxy> ref : looperThread.realm.handlerController.realmObjects.keySet()) {
+            assertFalse(ref.get() == allTypesPrimaryKey);
+        }
+
+        // Add different listeners twice
+        allTypesPrimaryKey.addChangeListener(new RealmChangeListener<RealmModel>() {
+            @Override
+            public void onChange(RealmModel element) {
+            }
+        });
+        allTypesPrimaryKey.addChangeListener(new RealmChangeListener<RealmModel>() {
+            @Override
+            public void onChange(RealmModel element) {
+            }
+        });
+
+        assertEquals(1, looperThread.realm.handlerController.realmObjects.size());
+        for (WeakReference<RealmObjectProxy> ref : looperThread.realm.handlerController.realmObjects.keySet()) {
+            assertTrue(ref.get() == allTypesPrimaryKey);
+        }
+        looperThread.testComplete();
+    }
+
+    // The object should not be added to HandlerController again after the async query loaded.
+    @Test
+    @RunTestInLooperThread
+    public void addChangeListener_checkHandlerRealmObjectsWhenCallingOnAsyncObject() {
+        looperThread.realm.beginTransaction();
+        looperThread.realm.createObject(AllTypesPrimaryKey.class, 1);
+        looperThread.realm.commitTransaction();
+
+        final AllTypesPrimaryKey allTypesPrimaryKey = looperThread.realm.where(AllTypesPrimaryKey.class).findFirstAsync();
+        allTypesPrimaryKey.addChangeListener(new RealmChangeListener<RealmModel>() {
+            @Override
+            public void onChange(RealmModel element) {
+                allTypesPrimaryKey.addChangeListener(new RealmChangeListener<RealmModel>() {
+                    @Override
+                    public void onChange(RealmModel element) {
+
+                    }
+                });
+                assertEquals(1, looperThread.realm.handlerController.realmObjects.size());
+                looperThread.testComplete();
+            }
+        });
+        assertEquals(1, looperThread.realm.handlerController.realmObjects.size());
+        for (RealmQuery<? extends RealmModel> query : looperThread.realm.handlerController.realmObjects.values()) {
+            assertNotNull(query);
         }
     }
 }